requirements.txt
discord.py
google-genai
cachetools
python-dotenv